ENH: don't reproduce streams class
